+++ src/main/scala/edu/ie3/simona/model/participant/HpModel.scala
@@ -115,20 +115,23 @@ final case class HpModel(
     * function calculates the heat pump's next state to get the actual active
     * power of this state use [[calculateActivePower]] with the generated state
     *
-    * @param state
-    *   Current state of the heat pump
+    * @param lastState
+    *   Last state of the heat pump
     * @param relevantData
     *   data of heat pump including
     * @return
-    *   next [[HpState]]
+    *   Booleans if Hp can operate and can be out of operation plus next
+    *   [[HpState]]
     */
   def determineState(
-      state: HpState,
+      lastState: HpState,
       relevantData: HpRelevantData,
-  ): HpState = {
-    val (turnOn, houseDemand, storageDemand) =
-      operatesInNextState(state, relevantData)
-    calcState(state, relevantData, turnOn, houseDemand, storageDemand)
+  ): (Boolean, Boolean, HpState) = {
+    val (turnOn, canOperate, canBeOutOfOperation, houseDemand, storageDemand) =
+      operatesInNextState(lastState, relevantData)
+    val updatedState =
+      calcState(lastState, relevantData, turnOn, houseDemand, storageDemand)
+    (canOperate, canBeOutOfOperation, updatedState)
   }
 
   /** Depending on the input, this function decides whether the heat pump will
@@ -142,8 +145,8 @@ final case class HpModel(
     * @param relevantData
     *   Relevant (external) data
     * @return
-    *   boolean defining if heat pump runs in next time step and if house and
-    *   storage have some demand
+    *   boolean defining if heat pump runs in next time step, if it can be in
+    *   operation and out of operation plus the demand of house and storage
     */
   private def operatesInNextState(
       state: HpState,
@@ -162,6 +165,7 @@ final case class HpModel(
         _.storedEnergy =~ zeroKWH
       )
 
+    //FIXME harmonise this
     val houseDemand =
       (demandHouse.hasRequiredDemand && noThermalStorageOrThermalStorageIsEmpty) || (state.isRunning && demandHouse.hasAdditionalDemand)
     val heatStorageDemand =
@@ -170,10 +174,18 @@ final case class HpModel(
     val turnHpOn: Boolean =
       houseDemand || heatStorageDemand
 
+    val canOperate =
+      demandHouse.hasRequiredDemand || demandHouse.hasAdditionalDemand ||
+        demandStorage.hasRequiredDemand || demandStorage.hasAdditionalDemand
+    val canBeOutOfOperation =
+      !(demandHouse.hasRequiredDemand && noStorageOrStorageIsEmpty)
+
     (
       turnHpOn,
-      houseDemand,
-      heatStorageDemand,
+      canOperate,
+      canBeOutOfOperation,
+      demandHouse,
+      demandStorage,
     )
   }
 
@@ -241,25 +253,8 @@ final case class HpModel(
       lastState: HpState,
   ): ProvideFlexOptions = {
     /* Determine the operating state in the given tick */
-    val updatedHpState: HpState = determineState(lastState, data)
-
-    /* Determine the options we have */
-    val (
-      thermalEnergyDemandHouse,
-      thermalEnergyDemandStorage,
-      updatedThermalGridState,
-    ) =
-      thermalGrid.energyDemandAndUpdatedState(
-        data.currentTick,
-        lastState.ambientTemperature.getOrElse(data.ambientTemperature),
-        data.ambientTemperature,
-        lastState.thermalGridState,
-      )
-    val canOperate =
-      thermalEnergyDemandHouse.hasRequiredDemand || thermalEnergyDemandHouse.hasAdditionalDemand ||
-        thermalEnergyDemandStorage.hasRequiredDemand || thermalEnergyDemandStorage.hasAdditionalDemand
-    val canBeOutOfOperation =
-      !thermalEnergyDemandHouse.hasRequiredDemand && !thermalEnergyDemandStorage.hasRequiredDemand
+    val (canOperate, canBeOutOfOperation, updatedHpState)
+        : (Boolean, Boolean, HpState) = determineState(lastState, data)
 
     val lowerBoundary =
       if (canBeOutOfOperation)