+++ src/main/scala/edu/ie3/simona/model/thermal/ThermalGrid.scala
@@ -13,6 +13,7 @@ import edu.ie3.datamodel.models.result.thermal.{
   CylindricalStorageResult,
   ThermalHouseResult,
 }
+import edu.ie3.simona.exceptions.InvalidParameterException
 import edu.ie3.simona.exceptions.agent.InconsistentStateException
 import edu.ie3.simona.model.thermal.ThermalGrid.{
   ThermalEnergyDemand,
@@ -103,8 +104,8 @@ final case class ThermalGrid(
       storage
         .zip(state.storageState)
         .map { case (storage, state) =>
-          val updatedStorageState =
-            storage.updateState(tick, state.qDot, state)._1
+          val (updatedStorageState, _) =
+            storage.updateState(tick, state.qDot, state)
           val storedEnergy = updatedStorageState.storedEnergy
           val soc = storedEnergy / storage.getMaxEnergyThreshold
           val storageRequired = {
@@ -785,13 +786,12 @@ object ThermalGrid {
 
     def hasRequiredDemand: Boolean = required > zeroMWH
 
-    def hasAdditionalDemand: Boolean = possible > required
+    def hasAdditionalDemand: Boolean = possible > zeroMWH
   }
   object ThermalEnergyDemand {
 
     /** Builds a new instance of [[ThermalEnergyDemand]]. If the possible energy
-      * is less than the required energy, this is considered to be a bad state
-      * and the required energy is curtailed to the possible energy.
+      * is less than the required energy, this is considered to be a bad state.
       * @param required
       *   The absolutely required energy to reach target state
       * @param possible
@@ -803,8 +803,12 @@ object ThermalGrid {
         required: Energy,
         possible: Energy,
     ): ThermalEnergyDemand = {
-      if (possible < required)
-        new ThermalEnergyDemand(possible, possible)
+      if (
+        math.abs(possible.toKilowattHours) < math.abs(required.toKilowattHours)
+      )
+        throw new InvalidParameterException(
+          s"The possible amount of energy {$possible} is smaller than the required amount of energy {$required}. This is not supported."
+        )
       else
         new ThermalEnergyDemand(required, possible)
     }